src/test/java/algoexpert/binary_tree/iterativeInOrderTraversal/Program.java
package algoexpert.binary_tree.iterativeInOrderTraversal;

import java.util.function.Function;

public class Program {
    public static void iterativeInOrderTraversal(BinaryTree tree, Function<BinaryTree, Void> callback) {
        BinaryTree currentNode = tree;
        BinaryTree prevNode = null;
        while (currentNode != null) {
            BinaryTree nextNode = null;
            if (prevNode == null || currentNode.parent == prevNode) {
                if (currentNode.left != null) {
                    nextNode = currentNode.left;
                } else {
                    callback.apply(currentNode);
                    nextNode = currentNode.right != null ? currentNode.right : currentNode.parent;
                }
            } else if (currentNode.left == prevNode) {
                callback.apply(currentNode);
                nextNode = currentNode.right != null ? currentNode.right : currentNode.parent;
            } else {
                nextNode = currentNode.parent;
            }
            prevNode = currentNode;
            currentNode = nextNode;
        }
    }

    static class BinaryTree {
        public int value;
        public BinaryTree left;
        public BinaryTree right;
        public BinaryTree parent;

        public BinaryTree(int value) {
            this.value = value;
        }

        public BinaryTree(int value, BinaryTree parent) {
            this.value = value;
            this.parent = parent;
        }
    }
}
